Library/requirement.txt
ultralytics
easyocr
cv2
numpy
tqdm
pyyaml
torch
matplotlib.pyplot